Updates to file structure
