+++ DJI-telemetry-log-parser.code-workspace
@@ -1,0 +1,11 @@
+{
+	"folders": [
+		{
+			"path": "../../../../../Users/edancain/DJI_Telemetry_parsing_multiple_architectures/DJI-telemetry-log-parser"
+		},
+		{
+			"path": "."
+		}
+	],
+	"settings": {}
+}
