+++ Dockerfile.aarch64-unknown-linux-gnu
@@ -1,0 +1,33 @@
+FROM ubuntu:latest
+
+RUN apt-get update && apt-get install -y \
+    build-essential \
+    gcc \
+    g++ \
+    gcc-aarch64-linux-gnu \
+    g++-aarch64-linux-gnu \
+    libc6-dev-arm64-cross \
+    curl \
+    pkg-config \
+    && rm -rf /var/lib/apt/lists/*
+
+RUN curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
+ENV PATH="/root/.cargo/bin:${PATH}"
+
+RUN rustup target add aarch64-unknown-linux-gnu
+
+RUN mkdir -p /root/.cargo && \
+    echo '[target.aarch64-unknown-linux-gnu]' >> /root/.cargo/config && \
+    echo 'linker = "aarch64-linux-gnu-gcc"' >> /root/.cargo/config
+
+ENV CC_aarch64_unknown_linux_gnu=aarch64-linux-gnu-gcc
+ENV CXX_aarch64_unknown_linux_gnu=aarch64-linux-gnu-g++
+
+WORKDIR /usr/src/myapp
+
+CMD cp -r /usr/src/myapp /tmp/build && \
+    cd /tmp/build && \
+    cargo build --release --target aarch64-unknown-linux-gnu && \
+    mkdir -p /usr/src/myapp/target/aarch64-unknown-linux-gnu/release && \
+    cp -r /tmp/build/target/aarch64-unknown-linux-gnu/release/* /usr/src/myapp/target/aarch64-unknown-linux-gnu/release/ && \
+    echo "Build completed for aarch64-unknown-linux-gnu"
