+++ README.md
@@ -1,5 +1,105 @@
 # dji-log-parser
 
+Edan Edit: 
+Original code found here:
+https://github.com/lvauvillier/dji-log-parser
+
+I extended this work and I provide an explanation of the automation process of creating a C-compatible interface of the RUST library for use by other languages.
+(nb: original implementation was manual creation of the header file. Now the process is automated as is explained below)
+
+The build.rs file works in conjunction with the c_api.rs file to create the dji-log-parser.h file. 
+    . build.rs: is a special RUST file that runs before the main compilation. It's purpose is to generate the C header file automatically.
+                : it determines the crate directory and package name.
+                : it sets up the output path for the header file. 
+                : it uses the cbindgen library to generate C bindings from the RUST code.
+
+The c_api.rs file contains the RUST functions that will be exposed to C.
+                : functions marked with #[no_mangle] and pub extern "C" are made available to C.
+                : these function use C-compatible types (like c_char instead of RUST's str).
+
+
+How it works:
+    . when the code is run with "cargo build", the build.rs script is executed first.
+    . it looks for functions marked with #[no_mangle] and pub extern "C".
+    . for each of these functions, it generates a corresponding C function declaration in the header file.
+
+Resulting dji-log-parser.h file.
+                ; this file is the interface that other languages use to interact with the RUST library.
+                : it declares the functions that are implemented in RUST but callable from C.
+
+
+The build.rs automates the process of creating a C-compatible interface for the RUST library. It ensures that the header
+file always matches the actual implementation in the RUST code, in order to reduce the chance of compile error from manual 
+updates.
+
+To use the RUST library ("libdji_log_parser.a") from other languages, include the generated dji-log-parser.h and link against library. 
+Golang example:
+
+    package mypackage
+
+    /*
+    #cgo LDFLAGS: -L${SRCDIR}/../target/release -ldji_log_parser
+    #cgo CFLAGS: -I${SRCDIR}/../dji-log-parser/include
+    #include "dji-log-parser.h"
+    #include <stdlib.h>
+    */
+    
+    import "C"
+    
+    <further imports here>
+    type GeoJSON struct {
+    Type     string    `json:"type"`
+    Features []Feature `json:"features"`
+    }
+    
+    type Feature struct {
+        Type       string     `json:"type"`
+        Geometry   Geometry   `json:"geometry"`
+        Properties Properties `json:"properties"`
+    }
+    
+    type Geometry struct {
+        Type        string    `json:"type"`
+        Coordinates []float64 `json:"coordinates"`
+    }
+    
+    type Properties struct {
+        Time   string  `json:"time"`
+        Height float64 `json:"height"`
+        Speed  float64 `json:"speed"`
+    }
+    
+    func processReader(reader io.Reader, apiKey string) (*geom.Geometry, error) {
+        data, err := io.ReadAll(reader)
+        if err != nil {
+            return nil, fmt.Errorf("error reading data: %s", err)
+        }
+    
+        cData := C.CBytes(data)
+        defer C.free(unsafe.Pointer(cData))
+        cLength := C.size_t(len(data))
+        cApiKey := C.CString(apiKey)
+        defer C.free(unsafe.Pointer(cApiKey))
+    
+        geojsonPtr := C.get_geojson_string_from_bytes((*C.uchar)(unsafe.Pointer(cData)), cLength, cApiKey)
+        if geojsonPtr == nil {
+            errPtr := C.get_last_error()
+            errStr := C.GoString(errPtr)
+            C.c_api_free_string(errPtr)
+            return nil, fmt.Errorf("failed to get GeoJSON: %s", errStr)
+        }
+        defer C.c_api_free_string(geojsonPtr)
+    
+        geojsonStr := C.GoString(geojsonPtr)
+    
+        var geojson GeoJSON
+        err = json.Unmarshal([]byte(geojsonStr), &geojson)
+        if err != nil {
+            return nil, fmt.Errorf("error parsing GeoJSON: %s", err)
+        }
+    }
+Finish Edit
+
 [![crates](https://img.shields.io/crates/v/dji-log-parser.svg)](https://crates.io/crates/dji-log-parser)
 [![docs.rs](https://docs.rs/dji-log-parser/badge.svg)](https://docs.rs/dji-log-parser)
