+++ dji-log-parser/Cargo.toml
@@ -2,12 +2,13 @@
 name = "dji-log-parser"
 version.workspace = true
 authors.workspace = true
-edition.workspace = true
+edition = "2021"
 description = "Library for parsing DJI txt logs"
 repository.workspace = true
 license.workspace = true
 keywords.workspace = true
 categories.workspace = true
+rust-version = "1.56"
 
 [dependencies]
 aes.workspace = true
@@ -23,6 +24,7 @@ ureq = { workspace = true, features = ["json"] }
 log = "0.4"
 geojson.workspace = true
 once_cell = "1.8.0"
+libc = "0.2"
 
 [lib]
 name = "dji_log_parser"
@@ -30,3 +32,10 @@ crate-type = ["staticlib", "rlib"]
 
 [build-dependencies]
 cbindgen = "0.24.0"
+cc = "1.0"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+openssl = { version = "0.10", features = ["vendored"] }
+
+[target.'cfg(target_os = "macos")'.build-dependencies]
+cc = "1.0"
