+++ dji-log-parser/build.rs
@@ -2,6 +2,9 @@ use std::env;
 use std::path::PathBuf;
 
 fn main() {
+    if cfg!(target_os = "linux") {
+        println!("cargo:rustc-link-lib=m");
+    }
     let crate_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
     let package_name = env::var("CARGO_PKG_NAME").unwrap();
     let output_file = PathBuf::from(&crate_dir)
