+++ docker-compose.yml
@@ -1,0 +1,23 @@
+version: '3.8'
+
+services:
+  x86_64-unknown-linux-gnu:
+    build:
+      context: .
+      dockerfile: Dockerfile.x86_64-unknown-linux-gnu
+    volumes:
+      - ./:/usr/src/myapp
+
+  aarch64-unknown-linux-gnu:
+    build:
+      context: .
+      dockerfile: Dockerfile.aarch64-unknown-linux-gnu
+    volumes:
+      - ./:/usr/src/myapp
+
+  linux-arm64-v8:
+    build:
+      context: .
+      dockerfile: Dockerfile.linux-arm64-v8
+    volumes:
+      - ./:/usr/src/myapp
